backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.0
